requirements.txt
africastalking==1.2.7
anyio==3.6.2
ariadne==0.21
ariadne-django==0.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
authentication==1.1.0
Authlib==1.3.1
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
dj-database-url==2.2.0
Django==5.0
django-auth==0.1.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.2
exceptiongroup==1.2.0
Flask==3.0.3
graphql-core==3.2.3
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
josepy==1.14.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
mozilla-django-oidc==4.0.1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyOpenSSL==24.0.0
pytest==8.2.2
python-dotenv==1.0.1
PyYAML==6.0.2rc1
redis==5.0.6
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
schema==0.7.7
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.34.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
whitenoise==6.6.0
WTForms==3.1.2
zope.interface==6.2
